refactor: extract ChannelMessageSender concern for multi-channel messaging
